Fix a int vs. size_t warning (unsigned int comparison)
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@1439 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ ReferenceAnalyser/referenceanalyser.cpp
@@ -324,7 +324,7 @@ void ReferenceAnalyser::PrintResult()
 
 	wxString result;
 	wxStringGraph::ConstNodeVectorRef path = m_classMembersGraph.getPath();
-	for (int i = 0; i < path.size(); i++) {
+	for (size_t i = 0; i < path.size(); i++) {
 		wxString node = path[i];
 
 		result += node;
